requirements.txt
pandas==3.0.1
numpy==2.4.2
sentence-transformers==5.2.3
faiss-cpu==1.13.2
google-genai==1.66.0
gradio==6.9.0
